Added rudimentary file watcher
Implemented a simple file watcher using 'watcher' package as 'watch'.

watch takes in the directories to watch (recursively, if provided) and a
regex filter to monitor file names matching the filter. It also takes in
a handler function that will be called whenever an event occurs on one
of the watched files. It currently watches only for 'write' events.

Also, added a basic test for fileops. This test file simply has helper
functions to create/delete/write/rename files. This will be used for
writing tests for 'watch' package.